rename of program to slurmCheck
